Started writing DataModalities object for storing data and meta information.
Different and potentially missing data modalities means that an object oriented approach is more sensible for storing the dataset for training and testing than a list or numpy array variable in Python due to potentially missing columns and the strong dependence on column order. An object approach could come with modality names and methods for creating training and test set arrays in a way that makes it easier to make changes to the structure/order in the future.

The number of features for each modality as attributes and other metadata info can easily be incorperated.