Implement DataModalities object with each point a DataPoint object
After much thought I decided to use one "superstructure class" for data
modalities and one object per data point. This is intended for training,
validation, and test data, and will not be used for inference.

Should in the long run make it easier to generate, save and load
training and test datasets.

Code is not tested yet.

GIT - Will from now on try to follow this convention:
https://chris.beams.io/posts/git-commit/#seven-rules