Fix DataModalities.add_meta() so that the metadata name input is used
Still need to fix so that a single input is handled without wrapping it in
a list in the input argument (do this in object function).

Also added soft_update() function to DataPoint class so that only the points
that get new data are updated. The rest get an empty field with a default
None/Nan value if it is a new metadata type. Otherwise, previously written
data is left intact.