Fix bug where different number of modality bands caused error
This was because np.asarray required the size of the lists in the
top list to be equal and converted it to a 3D array.
I.E. Output was a list with 165 lists, each with to lists with
a number of float equal to the number of bands (i.e. 3 and 10).
This was fine when both were equal to 3, but crashed when the
number of floats were not equal.

Fix: Modified how the modalities were read. They are now added to
a list instead of appending (the list) to the output list.