Fix bug caused by make_list and length functions
The fix was simply wrap input to DataModalities.add_to_point() in
list brackets, [], to circumvent the make_list function.
The bug was caused by a check in make_list which calls numel which
in turn calls length. The length function does not handle the
Pandas Timestamp object and tries to use len() on it, which fails.

Possible fixes in length includes checking for Pandas datatypes,
or using the fact that Timestamp.size fails for a single Timestamp
and not for series of Timestamps, and vice versa with
Timestamp.value, which works for a single input.

Could also consider removing the make_list function all togheter,
because right now it only wraps the input in brackets and throws
errors if input is multi dimensional.