Read in situ tree measurements and store in corresponding point
Need to handle identical column names (ID) for tree and vegetation
files. This is done by using the header from the vegetation file
as an 'exclusion list' parameter in the DataModalities.add_tree
function.

Initialization and list appending is handled in the
DataPoint.tree_update() function which is called for every column-
value pair of the data file (not in the exclusion list).