Add plots of LAI
TODO: Consider where derived variables, like LAI, average tree
height, etc., should be caluclated, and if they should be stored
in the data object.
Should probably either be calculated and stored when the data is
first added to the object, or calculated (and not stored) when the
data is read and plottet/analysed.