Remove old code for norm and file dialouge, add optical data once
Normalization was not used, normalized data could potentially be
added as a separate data type.

The "prompt user for file" dialouge is not necessary with the
current setup of looping over all datasets in list.

Many datasets (from the same date) contains the same optical data,
now optical data is only added for dataset(s) defined in a list.