Add dialog to choose between creating vs. adding to existing object
Uses a tkninter yes/no-question, but the dialog could be made
clearer.
Also, there should be three distinct choises:
1. Create new object.
2. Overwrite current object.
3. Add to current object.
Currently 1. and 2. correspond to the same answer in the dialog,
and is determined by the object name written in the script.
TODO: Add logic to handle this?

TODO: is root.destroy() needed?